Commit changes
need to commit and push to test functionality.

All of this stuff is still in flux, esp. since the materials of the site
change all the time with the new materials for each chapter, hence the
lack of constant commits, but alas, here we are now.

The plugin I am using to check layout doesnt render any of the images
and now I want to see if its working.